refactor command module
