+++ app/lib/cache_service/commands/cache_query_and_page_matches.rb
@@ -1,5 +1,5 @@
 module CacheService::Commands
-  class CacheQueryAndResults < Command::Base::Abstract
+  class CacheQueryAndPageMatches < Command::Abstract
     def initialize(query)
       super()
       @query = query
