+++ app/lib/parse_service/commands/insert_links.rb
@@ -1,5 +1,5 @@
 module ParseService::Commands
-  class InsertLinks < Command::Base::Abstract
+  class InsertLinks < Command::Abstract
     def initialize(page, parsed_page)
       super()
       @page = page
