+++ app/lib/parse_service/commands/upload_parsed_page_to_s3.rb
@@ -1,5 +1,5 @@
 module ParseService::Commands
-  class UploadParsedPageToS3 < Command::Base::Abstract
+  class UploadParsedPageToS3 < Command::Abstract
     def initialize(url, parsed_page)
       super()
       @url = url
